+++ testbed/src/Scene.cpp
@@ -33,7 +33,8 @@ using namespace openglframework;
 Scene::Scene(const std::string& name, EngineSettings& engineSettings, bool isShadowMappingEnabled)
       : mName(name), mEngineSettings(engineSettings), mLastMouseX(0), mLastMouseY(0), mInterpolationFactor(0.0f), mViewportX(0), mViewportY(0),
         mViewportWidth(0), mViewportHeight(0), mIsShadowMappingEnabled(isShadowMappingEnabled),
-        mIsContactPointsDisplayed(true), mIsAABBsDisplayed(false), mIsWireframeEnabled(false) {
+        mAreContactPointsDisplayed(true), mAreContactNormalsDisplayed(false), mAreBroadPhaseAABBsDisplayed(false),
+        mAreCollidersAABBsDisplayed(false), mAreCollisionShapesDisplayed(false), mIsWireframeEnabled(false) {
 
 }
 
@@ -184,30 +185,4 @@ void Scene::rotate(int xMouse, int yMouse) {
             }
         }
     }
-}
-
-// Called when some contacts occur
-void Scene::onContact(const rp3d::CollisionCallback::CallbackData& callbackData) {
-
-    // For each contact pair
-    for (uint p=0; p < callbackData.getNbContactPairs(); p++) {
-
-        rp3d::CollisionCallback::ContactPair contactPair = callbackData.getContactPair(p);
-
-        // For each contact point of the contact pair
-        for (uint c=0; c < contactPair.getNbContactPoints(); c++) {
-
-            rp3d::CollisionCallback::ContactPoint contactPoint = contactPair.getContactPoint(c);
-
-            rp3d::Vector3 point = contactPair.getCollider1()->getLocalToWorldTransform() * contactPoint.getLocalPointOnShape1();
-            rp3d::Vector3 normalWorld = contactPoint.getWorldNormal();
-            openglframework::Vector3 normal = openglframework::Vector3(normalWorld.x, normalWorld.y, normalWorld.z);
-            SceneContactPoint contact(openglframework::Vector3(point.x, point.y, point.z), normal, openglframework::Color::red());
-            mContactPoints.push_back(contact);
-        }
-    }
-}
-
-void Scene::onTrigger(const rp3d::OverlapCallback::CallbackData& callbackData) {
-
 }